Adds user profile management and enhances tag functionality
- Introduces a user profile feature with view and edit capabilities, including a new route and UI updates.
- Enhances the tag management system by adding list, update, and delete views with corresponding routes.
- Updates models to use UUIDs for IDs and indexes the `created_at` field for better query performance.
- Removes unused forms, views, and models, simplifying the codebase.
- Adds `django-widget-tweaks` for improved form rendering.

- Fixes inconsistencies in tag-related routes and templates.
- Removes redundant tests and updates templates to reflect the new features and structure.

Relates to #123 for feature enhancements and cleanup.
