+++ app/authentication/forms/__init__.py
@@ -1,0 +1,9 @@
+from .authentication_form import CustomAuthenticationForm
+from .user_creation_view import CustomUserCreationForm
+from .profile_form import ProfileForm
+
+__all__ = [
+    "CustomAuthenticationForm",
+    "CustomUserCreationForm",
+    "ProfileForm",
+]
