+++ app/common/models.py
@@ -9,7 +9,7 @@ class AbstractBaseModel(models.Model):
 
     id = models.UUIDField(primary_key=True, default=uuid4, editable=False, unique=True)
 
-    created_at = models.DateTimeField(auto_now_add=True)
+    created_at = models.DateTimeField(auto_now_add=True, db_index=True)
     updated_at = models.DateTimeField(auto_now=True)
 
     class Meta:
