+++ app/journal/models/__init__.py
@@ -1,0 +1,7 @@
+from .journal_entry_model import JournalEntry
+from .tag_model import Tag
+
+__all__ = [
+    "JournalEntry",
+    "Tag"
+]
