+++ app/journal/urls.py
@@ -6,6 +6,9 @@ from .views import (
     JournalListView,
     JournalDetailView,
     TagAutocompleteView,
+    TagListView,
+    TagUpdateView,
+    TagDeleteView
 )
 
 app_name = "journal"
@@ -17,4 +20,7 @@ urlpatterns = [
     path("entry/j/<uuid:entry_id>/", JournalDetailView.as_view(), name="entry_detail"),
     path("entry/e/<uuid:entry_id>/", EditJournalView.as_view(), name="edit_entry"),
     path("tags/autocomplete/", TagAutocompleteView.as_view(), name="tag_autocomplete"),
+    path("tags/list/", TagListView.as_view(), name="tag_list"),
+    path("tags/update/<str:pk>/", TagUpdateView.as_view(), name="tag_update"),
+    path("tags/delete/<str:pk>/", TagDeleteView.as_view(), name="tag_delete"),
 ]
