+++ app/templates/base.html
@@ -37,7 +37,7 @@
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/tags/">
+                            <a class="nav-link" href="{% url 'journal:tag_list' %}">
                                 <i class="bi bi-tags me-1"></i>Tags
                             </a>
                         </li>
@@ -51,6 +51,11 @@
                                 <i class="bi bi-person-circle me-1"></i>{{ user.username }}
                             </a>
                             <ul class="dropdown-menu">
+                                <li>
+                                    <a href="{% url 'authentication:profile' %}" class="dropdown-item">
+                                        <i class="bi bi-person me-1"></i>Profile
+                                    </a>
+                                </li>
                                 <li>
                                     <a href="{% url 'authentication:logout' %}" class="dropdown-item">
                                         <i class="bi bi-box-arrow-right me-1"></i>Logout
@@ -99,15 +104,6 @@
     <main class="container mt-4">
         {% block content %}{% endblock %}
     </main>
-
-    <!-- Footer -->
-    <footer class="bg-light py-4 mt-auto w-100">
-        <div class="container text-center">
-            <p class="text-muted mb-0">
-                &copy; {% now "Y" %} Daylog. Copyright &amp; all rights reserved.
-            </p>
-        </div>
-    </footer>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="{% static 'js/scripts.js' %}"></script>
