+++ app/templates/journal/tag_list.html
@@ -1,0 +1,45 @@
+{% extends "base.html" %}
+
+{% block title %}Tags - Daylog{% endblock %}
+
+{% block content %}
+<div class="row justify-content-center">
+    <div class="col-md-8">
+        <div class="card shadow">
+            <div class="card-body">
+                {% comment %} <div class="d-flex justify-content-between align-items-center mb-4">
+                    <h2 class="mb-0"><i class="bi bi-tags-fill text-primary me-2"></i>Your Tags</h2>
+                    <a href="{% url 'journal:tag_create' %}" class="btn btn-sm btn-success">
+                        <i class="bi bi-plus-circle me-1"></i>New Tag
+                    </a>
+                </div> {% endcomment %}
+
+                {% if tags %}
+                    <ul class="list-group list-group-flush">
+                        {% for tag in tags %}
+                            <li class="list-group-item d-flex justify-content-between align-items-center">
+                                <span><i class="bi bi-tag me-2 text-secondary"></i>{{ tag.name }}</span>
+                                <div>
+                                    <a href="{% url 'journal:tag_update' tag.pk %}" 
+                                       class="btn btn-sm btn-outline-primary me-2">
+                                        <i class="bi bi-pencil"></i> Edit
+                                    </a>
+                                    <a href="{% url 'journal:tag_delete' tag.pk %}" 
+                                       class="btn btn-sm btn-outline-danger">
+                                        <i class="bi bi-trash"></i> Delete
+                                    </a>
+                                </div>
+                            </li>
+                        {% endfor %}
+                    </ul>
+                {% else %}
+                    <div class="alert alert-info text-center">
+                        <i class="bi bi-info-circle me-2"></i>No tags found. 
+                        <a href="{% url 'journal:tag_create' %}">Create one</a>.
+                    </div>
+                {% endif %}
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
