+++ requirements.txt
@@ -6,3 +6,4 @@ psycopg2==2.9.10
 drf-spectacular==0.28.0
 django-debug-toolbar==6.0.0
 flake8==7.3.0
+django-widget-tweaks==1.5.0
